requirements.txt
faiss-cpu==1.7.4
ollama
numpy==1.24.4
requests==2.31.0
